Rename `requestHandler` as `requestMap`
