--- src/middlewares/paramValidator/requestHandler.ts
@@ -1,7 +1,0 @@
-import definitionRequest from './requestHandlers/definitionRequest';
-import userRequest from './requestHandlers/userRequest';
-
-export default {
-  ...definitionRequest,
-  ...userRequest,
-}
