+++ src/middlewares/paramValidator/requestMap/index.ts
@@ -1,0 +1,7 @@
+import definitionRequest from './definitionRequest';
+import userRequest from './userRequest';
+
+export default {
+  ...definitionRequest,
+  ...userRequest,
+}
