+++ src/middlewares/paramValidator/requestMap/userRequest.ts
@@ -3,6 +3,7 @@ import * as ApiURL from '@constants/ApiUrl';
 import { requireNonNull, optional } from '@src/utils/objectUtils';
 import SignInUserParam from '@models/RequestParam/SignInUserParam';
 import SignUpUserParam from '@models/RequestParam/SignUpUserParam';
+import ErrorType from '@constants/ErrorType';
 
 export default {
   [ApiURL.SESSION_NEW]: {
